creating MessagingViewModel, removing messages tab, and putting ChatListView into home tab
