+++ Roomi/Roomi/Core/Home/HomeView.swift
@@ -11,6 +11,7 @@ struct HomeView: View {
     var body: some View {
         NavigationStack {
             Text("Welcome to Roomi!")
+            ChatListView()
         }
     }
 }
