+++ Roomi/Roomi/Core/Messaging/ChatListView.swift
@@ -10,15 +10,14 @@ import SwiftUI
 struct ChatListView: View {
     @EnvironmentObject var viewModel: AuthViewModel
 
-        var body: some View {
-            List(viewModel.userList.filter { $0.id != viewModel.userSession?.uid }) { user in
-                NavigationLink(destination: ChatView(recipientId: user.id)) {
-                    Text("Chat with \(user.name)")
-                }
+    var body: some View {
+        List(viewModel.userList.filter { $0.id != viewModel.userSession?.uid }) { user in
+            NavigationLink(destination: ChatView(viewModel: MessagingViewModel(recipientId: user.id))) {
+                Text("Chat with \(user.name)")
             }
-            .onAppear {
-                viewModel.getAllUsers()
-            }
-            .navigationTitle("Messages")
         }
+        .onAppear {
+            viewModel.getAllUsers()
+        }
+    }
 }
