+++ Roomi/Roomi/Core/Root/ContentView.swift
@@ -28,12 +28,6 @@ struct ContentView: View {
                         .tabItem {
                             Label("Profile", systemImage: "person.circle")
                         }
-                    
-                    NavigationView
-                    {
-                        ChatListView()
-                    }.tabItem
-                    {Label("Messages", systemImage: "message.fill")}
                 }
             }
             else {
@@ -43,7 +37,7 @@ struct ContentView: View {
                 else {
                     SignUpView()
                 }
-                            }
+            }
         }
     }
 }
